add a mixin for models with location fields
