+++ common/models/__init__.py
@@ -1,3 +1,4 @@
+from .location import LocationMixin  # noqa
 from .lockable import Lockable  # noqa
 from .prefetch_snooping import PrefetchSnoopingMixin  # noqa
 from .thumbnails import ModelWithThumbnails  # noqa
